refactor: 클래스의 역할별로 패키지 위치 조정
