+++ src/main/java/racingcar/parser/InputParser.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.parser;
 
 import java.util.Arrays;
 import java.util.List;
@@ -8,7 +8,7 @@ public class InputParser {
 
     public List<String> parseInput(String input){
         return Arrays.stream(input.split(","))
-                .map(String::trim)   // 각 문자열의 앞뒤 공백을 제거
-                .collect(Collectors.toList());  // 결과를 리스트로 수집
+                .map(String::trim)
+                .collect(Collectors.toList());
     }
 }
