+++ src/main/java/racingcar/service/CarManager.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.service;
+
+import racingcar.model.Car;
 
 import java.util.ArrayList;
 import java.util.List;
