+++ src/main/java/racingcar/service/RaceService.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.service;
 
 import java.util.List;
 import java.util.Map;
